Run class (#2)
* Done. Segment picks best model out of presets.

* Main change: Segment class has parameter estimation, self test, and reporting. Now, only one report, bed, is implemented.

While waiting during tests, several other improvement are made all over the code.

* Just tiny things.

* Commit at the end of the day Apr 19th 2022

* Commit at the end of the day, Apr 21, 2022.

Half way done with solving windows in Run class. The combinations are redesign, so now Distribution returns combination of parameters that is used in solving as one of the solutions.

* Commit at the end of Apr 25th 2022

Run.py is implemented. Now the Chromosome.py needs to be redone to use Run.py

* Commit at the end of the day.

Code seems to work properly.

Nest is to write Segment.py so chromosome generates segments based on solutions from runs.

* Script runs!!

* attempting to shutdown logger

Co-authored-by: Dennis Kennetz <dennis.kennetz@stjude.org>